Refactor shared match helpers and reduce role-card fetches
- Extract duplicated payload and match helper logic into shared utils
- Reuse embedded role match data before falling back to vacant-role fetches
- Keep role badge hover expansion but require click to open role dropdown
