+++ src/components/teams/TeamCard.jsx
@@ -43,6 +43,16 @@ import { getMatchTier } from "../../utils/matchScoreUtils";
 import { getResultMatchScore } from "../../utils/teamMatchUtils";
 import { calculateDistanceKm } from "../../utils/locationUtils";
 import {
+  extractListPayload,
+  extractProfilePayload,
+} from "../../utils/payloadExtractors";
+import {
+  buildBadgeLookup,
+  buildTagLookup,
+  computeRoleUserMatch,
+  extractCandidateMatchData,
+} from "../../utils/matchHelpers";
+import {
   DEMO_ROLE_TOOLTIP,
   DEMO_TEAM_TOOLTIP,
   isSyntheticRole,
@@ -52,149 +62,13 @@ import DemoAvatarOverlay from "../users/DemoAvatarOverlay";
 
 const teamMemberBadgesCache = new Map();
 const viewerRoleProfileCache = new Map();
-const MATCH_WEIGHTS = {
-  tags: 0.4,
-  badges: 0.3,
-  distance: 0.3,
-};
-const LOCATION_GRACE_KM = 20;
-const LOCATION_GRACE_SCORE = 0.25;
+const EMPTY_ARRAY = [];
 
 const extractBadgeRows = (payload) => {
   if (Array.isArray(payload)) return payload;
   if (Array.isArray(payload?.data)) return payload.data;
   if (Array.isArray(payload?.data?.data)) return payload.data.data;
   return [];
-};
-
-const extractProfilePayload = (response) => {
-  const payload = response?.data ?? response;
-
-  if (!payload) return null;
-  if (payload?.success !== undefined) return payload?.data ?? null;
-
-  return payload?.data?.data ?? payload?.data ?? payload;
-};
-
-const extractListPayload = (response) => {
-  if (Array.isArray(response)) return response;
-  if (Array.isArray(response?.data)) return response.data;
-  if (Array.isArray(response?.data?.data)) return response.data.data;
-  return [];
-};
-
-const buildTagLookup = (tagData) => {
-  const nextMap = new Map();
-
-  for (const tag of tagData) {
-    nextMap.set(Number(tag.id), {
-      badgeCredits: Number(tag.badge_credits ?? tag.badgeCredits ?? 0),
-    });
-  }
-
-  return nextMap;
-};
-
-const buildBadgeLookup = (badgeData) => {
-  const nextMap = new Map();
-
-  for (const badge of badgeData) {
-    const name = (badge.badgeName ?? badge.badge_name ?? badge.name ?? "")
-      .trim()
-      .toLowerCase();
-    const credits = Number(
-      badge.totalCredits ?? badge.total_credits ?? badge.credits ?? 0,
-    );
-    const existing = nextMap.get(name);
-
-    nextMap.set(name, {
-      totalCredits: (existing?.totalCredits || 0) + credits,
-    });
-  }
-
-  return nextMap;
-};
-
-const roundMatchValue = (value) => Math.round(value * 100) / 100;
-
-const computeRoleUserMatch = ({
-  role,
-  tags,
-  badges,
-  user,
-  userTagMap,
-  userBadgeMap,
-}) => {
-  if (!role || !user) return null;
-
-  const requiredTagIds = tags
-    .map((tag) => Number(tag.tagId ?? tag.tag_id ?? tag.id))
-    .filter(Number.isFinite);
-  const requiredBadgeKeys = badges
-    .map((badge) =>
-      (badge.name ?? badge.badgeName ?? badge.badge_name ?? "")
-        .trim()
-        .toLowerCase(),
-    )
-    .filter(Boolean);
-
-  const matchingTags = requiredTagIds.filter((id) => userTagMap.has(id)).length;
-  const matchingBadges = requiredBadgeKeys.filter((key) => userBadgeMap.has(key)).length;
-
-  const tagScore =
-    requiredTagIds.length > 0 ? matchingTags / requiredTagIds.length : 0.5;
-  const badgeScore =
-    requiredBadgeKeys.length > 0
-      ? matchingBadges / requiredBadgeKeys.length
-      : 0.5;
-
-  const isRemote = role.isRemote ?? role.is_remote;
-  const maxDistanceKm = Number(role.maxDistanceKm ?? role.max_distance_km) || 50;
-
-  let distanceScore = 0.5;
-  let distanceKm = null;
-  let isWithinRange = null;
-
-  if (isRemote) {
-    distanceScore = 1;
-    isWithinRange = true;
-  } else {
-    distanceKm = calculateDistanceKm(user, role);
-
-    if (distanceKm !== null) {
-      if (distanceKm <= maxDistanceKm) {
-        distanceScore = 1;
-        isWithinRange = true;
-      } else if (distanceKm <= maxDistanceKm + LOCATION_GRACE_KM) {
-        distanceScore = LOCATION_GRACE_SCORE;
-        isWithinRange = false;
-      } else {
-        distanceScore = 0;
-        isWithinRange = false;
-      }
-    }
-  }
-
-  const matchScore =
-    MATCH_WEIGHTS.tags * tagScore +
-    MATCH_WEIGHTS.badges * badgeScore +
-    MATCH_WEIGHTS.distance * distanceScore;
-
-  return {
-    matchScore: roundMatchValue(matchScore),
-    matchDetails: {
-      tagScore: roundMatchValue(tagScore),
-      badgeScore: roundMatchValue(badgeScore),
-      distanceScore: roundMatchValue(distanceScore),
-      matchingTags,
-      totalRequiredTags: requiredTagIds.length,
-      matchingBadges,
-      totalRequiredBadges: requiredBadgeKeys.length,
-      distanceKm: distanceKm !== null ? Math.round(distanceKm) : null,
-      maxDistanceKm,
-      isWithinRange,
-    },
-  };
 };
 
 const getViewerRoleProfile = async (userId, fallbackUser = null) => {
@@ -255,6 +129,9 @@ const getExplicitMatchScore = (item) => {
 
   return Number.isFinite(score) ? score : null;
 };
+
+const hasRoleMatchDetails = (item) =>
+  item?.matchDetails != null || item?.match_details != null;
 
 const hasDisplayableBadges = (rawBadges) => {
   if (typeof rawBadges === "string") {
@@ -511,8 +388,12 @@ const TeamCard = ({
   const flatRoleBio = roleSource?.bio ?? roleSource?.roleBio ?? null;
   const flatRoleCity = roleSource?.city ?? null;
   const flatRoleCountry = roleSource?.country ?? null;
-  const flatRoleTags = roleSource?.tags ?? [];
-  const flatRoleBadges = roleSource?.badges ?? [];
+  const flatRoleTags = roleSource?.tags ?? EMPTY_ARRAY;
+  const flatRoleBadges = roleSource?.badges ?? EMPTY_ARRAY;
+  const roleHasPreloadedRequirements =
+    (Array.isArray(nestedRoleData?.tags) &&
+      Array.isArray(nestedRoleData?.badges)) ||
+    (Array.isArray(roleSource?.tags) && Array.isArray(roleSource?.badges));
   const flatRoleIsRemote =
     roleSource?.isRemote ?? roleSource?.is_remote ?? undefined;
   const syntheticRoleFlag = isRoleApplicationVariant
@@ -1250,21 +1131,28 @@ const TeamCard = ({
 
     const fetchRoleMatchData = async () => {
       try {
-        const [viewerProfileRes, detailsRes] = await Promise.allSettled([
-          getViewerRoleProfile(user.id, user),
-          vacantRoleService.getVacantRoleById(roleTeamId, roleDataId),
-        ]);
-
-        if (viewerProfileRes.status !== "fulfilled") {
-          throw viewerProfileRes.reason;
+        const preloadedMatchData = extractCandidateMatchData(roleData);
+        if (
+          preloadedMatchData.matchScore != null &&
+          hasRoleMatchDetails(roleData)
+        ) {
+          if (!isCancelled) {
+            setRoleMatchData(preloadedMatchData);
+          }
+          return;
         }
 
-        const viewerProfile = viewerProfileRes.value;
-        const hydratedRole =
-          detailsRes.status === "fulfilled"
-            ? extractProfilePayload(detailsRes.value)
-            : null;
-        const effectiveRole = hydratedRole ?? roleData ?? null;
+        const viewerProfile = await getViewerRoleProfile(user.id, user);
+        let effectiveRole = roleData ?? null;
+
+        if (!roleHasPreloadedRequirements) {
+          const detailsRes = await vacantRoleService.getVacantRoleById(
+            roleTeamId,
+            roleDataId,
+          );
+          effectiveRole = extractProfilePayload(detailsRes) ?? effectiveRole;
+        }
+
         const nextMatchData = effectiveRole
           ? computeRoleUserMatch({
               role: effectiveRole,
@@ -1289,7 +1177,15 @@ const TeamCard = ({
     return () => {
       isCancelled = true;
     };
-  }, [isRoleVariant, showMatchScore, roleDataId, roleTeamId, user?.id]);
+  }, [
+    isRoleVariant,
+    showMatchScore,
+    roleData,
+    roleDataId,
+    roleHasPreloadedRequirements,
+    roleTeamId,
+    user?.id,
+  ]);
 
   // ================= GUARD CLAUSE – AFTER ALL HOOKS =================
